drivers: i2c: support storing multiple reg addresses
Support storing multiple reg addresses with a new Kconfig symbol,
`I2C_FALLBACK_ADDRESSES`. It is relatively common for I2C devices to
be available in several variants with the only difference being the
I2C address. This change allows all the possible addresses to be
specified. This opens up the potential for boards to populate any
variant depending on component availability, without changing binaries.

Signed-off-by: Jordan Yates <jordan.yates@data61.csiro.au>