bluetooth: hci: spi: drop `HOST_NUM_COMPLETED_PACKETS` rsp
Drop `HOST_NUM_COMPLETED_PACKETS` responses before attempting to
allocate a buffer, as this unexpected response can cause corruption of
a pending command packet if received at the wrong moment. The response
is ignored anyway by the handlers (after corrupting the command), so
there are no undesired side-effects from dropping it here.

This is a temporary workaround for the SPI driver only until #64158 can
be solved in the HCI layer.

Signed-off-by: Jordan Yates <jordan.yates@data61.csiro.au>